requirements.txt
Django==5.0.1
django-constance==3.1.0
requests==2.31.0
